Moved help and whatsnew to profile section and stories updated
