+++ example/src/Playground.js
@@ -65,11 +65,6 @@ const Playground = () => {
           showAppSwitcher
           onAppSwitcherToggle={() => toggleAppSwitcher((open) => !open)}
           appName="neetoUI"
-          showChangelog
-          changelogProps={{
-            icon: Notification,
-            onClick: () => alert("onChangelogToggle"),
-          }}
           footerLinks={FOOTER_LINKS}
         />
         <div className="relative flex flex-col flex-grow h-screen overflow-auto">
