+++ lib/components/layouts/Sidebar/Footer/index.js
@@ -1,11 +1,10 @@
 import React, { useState } from "react";
 import { NavLink } from "react-router-dom";
 
-import ChangeLog from "./ChangeLog";
 import AppSwitcher from "./AppSwitcher";
 import NavIconWrapper from "../NavIconWrapper";
 import ProfileSection from "../ProfileSection";
-import ProfileSectionTooltip from "../ProfileSectionTooltip";
+import ProfileSectionTooltip from "../ProfileSection/Tooltip";
 import Tooltip from "../../../Tooltip";
 import Typography from "../../../Typography";
 
@@ -14,9 +13,7 @@ const Footer = ({
   footerLinks = [],
   profileInfo,
   tooltipStyle,
-  changelogProps,
   collapsed,
-  showChangelog = false,
   showAppSwitcher = false,
   onAppSwitcherToggle,
 }) => {
@@ -64,13 +61,6 @@ const Footer = ({
           </NavIconWrapper>
         );
       })}
-      {showChangelog && (
-        <ChangeLog
-          collapsed={collapsed}
-          tooltipStyle={tooltipStyle}
-          changelogProps={changelogProps}
-        />
-      )}
       {showAppSwitcher && (
         <AppSwitcher
           appName={appName}
