+++ lib/components/layouts/Sidebar/ProfileSection/index.js
@@ -1,8 +1,8 @@
 import React, { forwardRef } from "react";
 import { Down } from "@bigbinary/neeto-icons";
 
-import Typography from "../../Typography";
-import Avatar from "../../Avatar";
+import Typography from "../../../Typography";
+import Avatar from "../../../Avatar";
 
 const ProfileSection = forwardRef(
   ({ collapsed, profileInfo, onClick }, ref) => {
