+++ lib/components/layouts/Sidebar/constant.js
@@ -38,6 +38,16 @@ export const SIDEBAR_PROPTYPES = {
     name: PropTypes.string,
     email: PropTypes.string,
     customContent: PropTypes.element,
+    changelogProps: PropTypes.shape({
+      onClick: PropTypes.func,
+      label: PropTypes.string,
+      icon: PropTypes.element,
+    }),
+    helpProps: PropTypes.shape({
+      onClick: PropTypes.func,
+      label: PropTypes.string,
+      icon: PropTypes.element,
+    }),
     topLinks: PropTypes.arrayOf(
       PropTypes.shape({
         onClick: PropTypes.func,
@@ -54,12 +64,6 @@ export const SIDEBAR_PROPTYPES = {
     ),
   }),
   isCollapsed: PropTypes.bool,
-  showChangelog: PropTypes.bool,
-  changelogProps: PropTypes.shape({
-    id: PropTypes.string,
-    icon: PropTypes.element,
-    onClick: PropTypes.func,
-  }),
   showAppSwitcher: PropTypes.bool,
   onAppSwitcherToggle: PropTypes.func,
   appName: PropTypes.string,
