+++ lib/components/layouts/Sidebar/index.js
@@ -14,13 +14,10 @@ const Sidebar = ({
   footerLinks = [],
   profileInfo,
   isCollapsed = true,
-  showChangelog = false,
-  changelogProps = {},
   showAppSwitcher = false,
   onAppSwitcherToggle = () => {},
   appName,
 }) => {
-
   return (
     <motion.div
       animate={isCollapsed ? "collapsed" : "expanded"}
@@ -48,9 +45,7 @@ const Sidebar = ({
         profileInfo={profileInfo}
         collapsed={isCollapsed}
         tooltipStyle={tooltipStyle}
-        showChangelog={showChangelog}
         showAppSwitcher={showAppSwitcher}
-        changelogProps={changelogProps}
         onAppSwitcherToggle={onAppSwitcherToggle}
       />
     </motion.div>
